Copying files from my local git
I am copying the project files from my local machine to this git. My intent is to try to save the revision history of my project. In the case that the revision history isn't accessible, it can be seen on the repo for this project on my actual GitHub account: https://github.com/Tello-/Snake-in-Pygame